refactor(internal): gather models in one file
