+++ internal/internal.go
@@ -3,7 +3,6 @@ package internal
 import (
 	"encoding/json"
 	"os"
-	"time"
 
 	"github.com/spacemonkeygo/errors"
 )
@@ -12,22 +11,6 @@ var (
 	InvalidJSONError = errors.NewClass("InvalidJSONError")
 	OutputError      = errors.NewClass("OutputError")
 )
-
-type Source struct {
-	Host       string  `json:"host"`
-	Port       float64 `json:"port"`
-	User       string  `json:"user"`
-	Password   string  `json:"password"`
-	PrivateKey string  `json:"private_key"`
-}
-
-type Version struct {
-	Timestamp time.Time `json:"time"`
-}
-
-type Params struct {
-	Script string `json:"script"`
-}
 
 func NewRequestFromStdin(stdin *os.File, request *Request) error {
 	err := json.NewDecoder(stdin).Decode(&request)
